Added Toon Shader, GameBoard Pieces, CubeScript, GameScript, and ARCamera Script
CubeScript will be used to highlight the board piece when the user is hovering over it.
GameScript will be used to model the game mechanics.
ARCameraScript is a redacted version of Ramzi's script from Assignment3. It:
  - Focuses / Defocuses on objects
  - Selects them
  - Allows for translation

I've also setup the image target with a plane on top of it to signify the board. The plane will
be removed when we finish setting up the GameScript's initialization function. I have not loaded the
image target or camera with any image target data from Vuforia.

Currently the GameScript creates a single instance of the GameBoard Piece, and colors it green. It will
create an 8x8 game board upon instantiation, and every other GameBoard Piece will be colored black.